blof/apps/blof/api/authenticators.py
from rest_framework.authentication import BasicAuthentication


class QuietBasicAuthentication(BasicAuthentication):
    def authenticate_header(self, request):
        return 'xBasic realm="%s"' % self.www_authenticate_realm