blof/requirements/runtime.txt
Django==1.6
djangorestframework==2.3.11
South==0.8.2